refactor(tests): Split `test_main` into separate files.
Rename test_frameworks to frameworks.
