+++ src/dvclive/utils.py
@@ -176,3 +176,21 @@ def catch_and_warn(exception, logger, on_finally=None):
 def rel_path(path, dvc_root_path):
     absolute_path = Path(path).absolute()
     return str(Path(os.path.relpath(absolute_path, dvc_root_path)).as_posix())
+
+
+def read_history(live, metric):
+    from dvclive.plots.metric import Metric
+
+    history, _ = parse_metrics(live)
+    steps = []
+    values = []
+    name = os.path.join(live.plots_dir, Metric.subfolder, f"{metric}.tsv")
+    for e in history[name]:
+        steps.append(int(e["step"]))
+        values.append(float(e[metric]))
+    return steps, values
+
+
+def read_latest(live, metric_name):
+    _, latest = parse_metrics(live)
+    return latest["step"], latest[metric_name]
